move abort_on_drop test to tests/spawn_ready
